refactor: rename ViewInvalidationService to ViewOrchestrator and flatten infrastructure module
Pure mechanical refactor in fold_db:

- Rename ViewInvalidationService -> ViewOrchestrator. The type handles
  dependency-graph traversal, invalidation, topological ordering, and
  background precomputation -- "invalidation" undersold the scope. File
  view_invalidation.rs -> view_orchestrator.rs, field view_invalidation_service
  -> view_orchestrator.

- Flatten fold_db_core/infrastructure/ into fold_db_core/. Only 4 files
  remained after message_bus moved out; the module name was misleading.
  event_monitor.rs, event_statistics.rs, pending_task_tracker.rs, and
  process_results_subscriber.rs now live at the top of fold_db_core, and
  all `super::infrastructure::*` / `fold_db_core::infrastructure::*` imports
  were rewritten accordingly.

No behavior changes. cargo fmt, clippy -D warnings, and full test suite
all pass.
